requirements.txt
Flask==2.2.5
gunicorn==21.2.0
pyswisseph==2.10.2